refactor: CelebProfile 컴포넌트 생성 및 useOnClickBlock 적용
