+++ frontend/src/components/CelebProfile/index.tsx
@@ -1,0 +1,3 @@
+import CelebProfile from './CelebProfile';
+
+export default CelebProfile;
